Document Object

Document Object

Document(name = NULL, createTime = NULL, updateTime = NULL, fields = NULL)

Arguments

name
The resource name of the document, for example projects/{project_id}/databases/{database_id}/documents/{document_path}

createTime
Output only

updateTime
Output only

fields
The document's fields. The map keys represent field names. A simple field name contains only characters a to z, A to Z, 0 to 9, or _, and must not start with 0 to 9. For example, foo_bar_17. Field names matching the regular expression __.*__ are reserved. Reserved field names are forbidden except in certain documented contexts. The map keys, represented as UTF-8, must not exceed 1,500 bytes and cannot be empty. Field paths may be used in other contexts to refer to structured fields defined here. For map_value, the field path is represented by the simple or quoted field names of the containing fields, delimited by .. For example, the structured field 'foo' : { map_value: { 'x&y' : { string_value: 'hello' }}} would be represented by the field path foo.x&y. Within a field path, a quoted field name starts and ends with ` and may contain any character. Some characters, including `, must be escaped using a \. For example, `x&y` represents x&y and `bak\`tik` represents bak`tik.

Value

Document object

Details

Autogenerated via gar_create_api_objects A Firestore document. Must not exceed 1 MiB - 4 bytes.

See also