DocumentChange Object

DocumentChange Object

DocumentChange(targetIds = NULL, removedTargetIds = NULL, document = NULL)

Arguments

targetIds

A set of target IDs of targets that match this document

removedTargetIds

A set of target IDs for targets that no longer match this document

document

The new state of the Document

Value

DocumentChange object

Details

Autogenerated via gar_create_api_objects A Document has changed. May be the result of multiple writes, including deletes, that ultimately resulted in a new value for the Document. Multiple DocumentChange messages may be returned for the same logical change, if multiple targets are affected.