GoogleLongrunningOperation Object

GoogleLongrunningOperation Object

GoogleLongrunningOperation(
  metadata = NULL,
  response = NULL,
  done = NULL,
  error = NULL,
  name = NULL
)

Arguments

metadata

Service-specific metadata associated with the operation. It typically contains progress information and common metadata such as create time. Some services might not provide such metadata. Any method that returns a long-running operation should document the metadata type, if any.

response

The normal response of the operation in case of success. If the original method returns no data on success, such as Delete, the response is google.protobuf.Empty. If the original method is standard Get/Create/Update, the response should be the resource. For other methods, the response should have the type XxxResponse, where Xxx is the original method name. For example, if the original method name is TakeSnapshot(), the inferred response type is TakeSnapshotResponse.

done

If the value is false, it means the operation is still in progress

error

The error result of the operation in case of failure or cancellation

name

The server-assigned name, which is only unique within the same service that originally returns it

Value

GoogleLongrunningOperation object

Details

Autogenerated via gar_create_api_objects This resource represents a long-running operation that is the result of a network API call.