Status Object

Status Object

Status(details = NULL, message = NULL, code = NULL)

Arguments

details
A list of messages that carry the error details

message
A developer-facing error message, which should be in English

code
The status code, which should be an enum value of google

Value

Status object

Details

Autogenerated via gar_create_api_objects The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.